# swagger.yaml
openapi: 3.0.0
info:
  title: "REST API"
  version: "1.0.0"
  description: "Example docs"
servers:
  - url: "https://metaswap.freeblock.site"
  # - url: "http://localhost:3001"
    description: "Dev server"
components:
  securitySchemes:
    XAuthToken:
      type: apiKey
      in: header
      name: X-Auth-Token
      description: "Provide your X-Auth-Token for authorization"
    AdminAuthToken:
      type: apiKey
      in: header
      name: authorization
      description: "Provide your authorization token for admin authrization"
...
security:
  - XAuthToken: []
  - AdminAuthToken: []
paths:
...
  /api/v1/account/get-user:
    get:
      summary: Get user information
      description: Retrieve user information by user ID (Admin permission required).
      operationId: getUser
      tags:
        - Account(Admin)
      parameters:
        - in: query
          name: userId
          required: true
          schema:
            type: integer
            example: 1
      responses:
        '200':
          description: User information retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: integer
                    example: 1
                  publicId:
                    type: string
                    example: "36"
                  telegramUserId:
                    type: string
                    example: "13"
                  telegramUsername:
                    type: string
                    nullable: true
                  telegramPhoto:
                    type: string
                    nullable: true
                  telegramPhotoMime:
                    type: string
                    nullable: true
                  firstName:
                    type: string
                    example: "robin"
                  lastName:
                    type: string
                    nullable: true
                  language:
                    type: string
                    nullable: true
                  email:
                    type: string
                    nullable: true
                  priceCurrency:
                    type: string
                    nullable: true
                  totpSecret:
                    type: string
                    nullable: true
                  userRole:
                    type: string
                    example: "USER"
                  userPermission:
                    type: string
                    example: "UNBLOCK"
                  adminPermission:
                    type: string
                    example: "UNBLOCK"
                  createdAt:
                    type: string
                    format: date-time
                    example: "2024-09-07T14:17:48.697Z"
                  Asset:
                    type: array
                    items:
                      type: object
                  Transaction:
                    type: array
                    items:
                      type: object
                  internalTransferUserOutputs:
                    type: array
                    items:
                      type: object
                  internalTransferUserInputs:
                    type: array
                    items:
                      type: object
                  Autoconvert:
                    type: array
                    items:
                      type: object
                  verification:
                    type: object
                    nullable: true
                  verificationRequests:
                    type: array
                    items:
                      type: object
                  emailUpdateRequests:
                    type: array
                    items:
                      type: object
                  deposits:
                    type: array
                    items:
                      type: object
                  payouts:
                    type: array
                    items:
                      type: object
                  frozenAmounts:
                    type: array
                    items:
                      type: object
                  freezeTransactions:
                    type: array
                    items:
                      type: object
                  unfreezeTransactions:
                    type: array
                    items:
                      type: object
                  Swaps:
                    type: array
                    items:
                      type: object
                      properties:
                        id:
                          type: integer
                          example: 1
                        sourceCurrency:
                          type: string
                          example: "ETH"
                        destinationCurrency:
                          type: string
                          example: "USDT"
                        sourceAmount:
                          type: number
                          format: float
                          example: 1
                        destinationAmount:
                          type: number
                          format: float
                          example: 2400
                        userId:
                          type: integer
                          example: 1
                        hashId:
                          type: string
                          example: "134545353"
                        createdAt:
                          type: string
                          format: date-time
                          example: "2024-09-07T05:21:53.279Z"
                        state:
                          type: string
                          example: "IN_PROGRESS"
        '400':
          description: Invalid request (e.g., missing or invalid fields)
        '401':
          description: Unauthorized - Admin permission required
        '404':
          description: User not found
        '500':
          description: Internal Server Error
  /api/v1/account/update-user-permission:
    post:
      summary: Update user permission
      description: Update the permission status of a user (Admin permission required).
      operationId: updateUserPermission
      tags:
        - Account(Admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                userId:
                  type: integer
                  example: 1
                permission:
                  type: string
                  enum: ["BLOCK", "UNBLOCK"]
                  example: "BLOCK"
      responses:
        '200':
          description: User permission updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  kind:
                    type: string
                    example: "OK"
                  message:
                    type: string
                    example: "user permission updated"
                  user:
                    type: object
                    properties:
                      id:
                        type: integer
                        example: 1
                      publicId:
                        type: string
                        example: "36"
                      telegramUserId:
                        type: string
                        example: "13"
                      telegramUsername:
                        type: string
                        nullable: true
                        example: null
                      telegramPhoto:
                        type: string
                        nullable: true
                        example: null
                      telegramPhotoMime:
                        type: string
                        nullable: true
                        example: null
                      firstName:
                        type: string
                        example: "robin"
                      lastName:
                        type: string
                        nullable: true
                        example: null
                      language:
                        type: string
                        nullable: true
                        example: null
                      email:
                        type: string
                        nullable: true
                        example: null
                      priceCurrency:
                        type: string
                        nullable: true
                        example: null
                      totpSecret:
                        type: string
                        nullable: true
                        example: null
                      userRole:
                        type: string
                        example: "USER"
                      userPermission:
                        type: string
                        enum: ["UNBLOCK", "BLOCK"]
                        example: "UNBLOCK"
                      createdAt:
                        type: string
                        format: date-time
                        example: "2024-09-07T14:17:48.697Z"
        '400':
          description: Invalid request (e.g., missing required fields)
        '401':
          description: Unauthorized - Admin permission required
        '500':
          description: Internal Server Error
  /api/v1/account/update-admin-permission:
    post:
      summary: Update admin permission
      description: Update the admin permission status of a user (Super permission required).
      operationId: updateAdminPermission
      tags:
        - Account(Super)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                userId:
                  type: integer
                  example: 1
                permission:
                  type: string
                  enum: ["BLOCK", "UNBLOCK"]
                  example: "BLOCK"
      responses:
        '200':
          description: Admin permission updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  kind:
                    type: string
                    example: "OK"
                  message:
                    type: string
                    example: "admin permission updated"
                  user:
                    type: object
                    properties:
                      id:
                        type: integer
                        example: 1
                      publicId:
                        type: string
                        example: "36"
                      telegramUserId:
                        type: string
                        example: "13"
                      telegramUsername:
                        type: string
                        nullable: true
                        example: null
                      telegramPhoto:
                        type: string
                        nullable: true
                        example: null
                      telegramPhotoMime:
                        type: string
                        nullable: true
                        example: null
                      firstName:
                        type: string
                        example: "robin"
                      lastName:
                        type: string
                        nullable: true
                        example: null
                      language:
                        type: string
                        nullable: true
                        example: null
                      email:
                        type: string
                        nullable: true
                        example: null
                      priceCurrency:
                        type: string
                        nullable: true
                        example: null
                      totpSecret:
                        type: string
                        nullable: true
                        example: null
                      userRole:
                        type: string
                        example: "ADMIN"
                      userPermission:
                        type: string
                        enum: ["UNBLOCK", "BLOCK"]
                        example: "UNBLOCK"
                      adminPermission:
                        type: string
                        enum: ["UNBLOCK", "BLOCK"]
                        example: "BLOCK"
                      createdAt:
                        type: string
                        format: date-time
                        example: "2024-09-07T14:17:48.697Z"
        '400':
          description: Invalid request (e.g., missing required fields)
        '401':
          description: Unauthorized - Super permission required
        '500':
          description: Internal Server Error
  /api/v1/account/add-admin:
    post:
      summary: Add admin privileges to a user
      description: Grant admin privileges to a user (Super permission required).
      operationId: addAdmin
      tags:
        - Account(Super)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  example: admin@email.com
                password:
                  type: string
                  example: Admin#1552
                firstName:
                  type: string
                  example: Artem
                lastName:
                  type: string
                  example: webdev
      responses:
        '200':
          description: Successfully added admin privileges
          content:
            application/json:
              schema:
                type: object
                properties:
                  kind:
                    type: string
                    example: "OK"
                  message:
                    type: string
                    example: "add admin success"
                  user:
                    type: object
                    properties:
                      id:
                        type: integer
                        example: 1
                      publicId:
                        type: string
                        example: "36"
                      telegramUserId:
                        type: string
                        example: "13"
                      telegramUsername:
                        type: string
                        nullable: true
                        example: null
                      telegramPhoto:
                        type: string
                        nullable: true
                        example: null
                      telegramPhotoMime:
                        type: string
                        nullable: true
                        example: null
                      firstName:
                        type: string
                        example: "robin"
                      lastName:
                        type: string
                        nullable: true
                        example: null
                      language:
                        type: string
                        nullable: true
                        example: null
                      email:
                        type: string
                        nullable: true
                        example: null
                      priceCurrency:
                        type: string
                        nullable: true
                        example: null
                      totpSecret:
                        type: string
                        nullable: true
                        example: null
                      userRole:
                        type: string
                        example: "ADMIN"
                      userPermission:
                        type: string
                        enum: ["UNBLOCK", "BLOCK"]
                        example: "UNBLOCK"
                      adminPermission:
                        type: string
                        enum: ["UNBLOCK", "BLOCK"]
                        example: "UNBLOCK"
                      createdAt:
                        type: string
                        format: date-time
                        example: "2024-09-07T14:17:48.697Z"
        '400':
          description: Invalid request (e.g., missing required fields)
        '401':
          description: Unauthorized - Super permission required
        '500':
          description: Internal Server Error
...